Transitively link C++ libraries properly.
Prior to this change, the Rust rules have a conceptual mismatch with how
the C++ rules work. This makes writing Rust bindings to a native library
impossible.

Bazel's model of C++ linking is to underlink all intermediate cc_library
targets and do the final link at the cc_binary target. To bring the Rust
rules in line with the C++ rules, this change removes any native library
linking from rust_library targets, but provides the information
transitively so that the rust_binary target can link all transitive
native libraries during the final link.

One thing to note is that Rust libraries (.rlib) behave slightly
differently. While they do not contain their rlib dependencies, they do
reference their rlib dependencies, so the proper behavior with rlibs is
to always link (--extern) the direct rlib deps, regardless of whether we
are building a rust_library or rust_binary.

I added an example to the ffi examples of the following dep chain:
  rust_binary -> rust_library -> cc_library -> cc_library

This is the typical way of binding a C++ library, where the cc_library
on the far right is the actual C++ library you are binding and the one
upstream from it is a C API over the C++ library. The rust_library
upstream of that is the Rust FFI declarations and finally the
rust_binary calls into the bindings through the rust_library. For good
measure, I added an additional rust_library dependency to the
rust_library to make sure that we still do the right thing with respect
to rust_library -> rust_library dependencies.