Allow crate test configurations to allow for ‟dev-deps”
Due to the way unit tests are handled with the rules, it is tricky for
a crate containing internal test configuration to have additional (test
only) deps.

Consider some source code like such

```rust
fn some_fn() -> u32 {
    2
}

mod tests {
    use super::some_fn();

    // This following line is not going to work
    use pretty_assertions::assert_eq;

    #[test]
    fn test_is_2() {
        assert_eq!(1, some_fn());
    }
}
```

With a build function like so

```python
rust_library(
    name = "example",
    srcs = ["lib.rs"]
)

rust_test(
    name = "example_test",
    deps = [
        ":example",
        "//third_party/rust:pretty_assertion",
    ],
)
```

Presently, due to the handling of ‟unit tests” the extra deps will
conflict with rust_rules ideas about how to identify the test, and give
an error along the lines of

```
attribute srcs: No lib.rs or example_test.rs source file found.
```

With this change we make it possible for tests to express ‟dev
dependencies”. This lets us avoid adding deps that are strictly for
testing.

This is somewhat distinct from `testonly` attributes as it does not
always hold that a dependency is always and only ever `testonly` in all
library circumstances, consider for example the usage of
https://github.com/pingcap/fail-rs. There are instances where this needs
to be used both in source code *and* test code, and is not easily
seperated from a cargo-raze perspective.

Closes #202

WIP