Add cfg = proc_macro_host_transition to all rules
The rust_library rule specifies a cfg option.  This performs transitions
in order to support WASM.  Other rules, such as rust_binary, do not have
this configuration, which causes Bazel to build and link some libraries
into the final binary twice, resulting in compilation errors.  This is
detailed in pull request #294 and issue #300.  The issue was introduced
in commit 7cde1e46.

This adds the example from @colin353 in PR #294 as an example that would
otherwise fail.

Fixes #300