Update proto crates: protobuf to 2.8.2 and grpc-rust to 0.6.2
All of the files within //proto/raze/remote are automatically generated
by cargo raze.

This updates the protobuf crate to 2.8.2, which is as far as we are able
to with the most recent version of grpc-rust, which is 0.6.2.  The
downstream dependencies of grpc-rust required specifying an upper bound
on the log 0.4 version, which would otherwise break log 0.3 dependent
crates.  This should be revisited if grpc-rust is changed or reevaluated
for a different default library.

This also introduces a patch folder, for patches to be applied by cargo
raze.  The rationale for patching protobuf is explained in the README,
but, in short, circumvents the build.rs file.