Add the ability to pass environment variables to rustc
This patch enables callers to supply an additional
parameter to rust targets (such as "rust_library", "rust_binary",
etc):

  rustc_env = [
    "KEY=VALUE",
  ]

This parameter parses the key/value pairs, and appends them
to the list of environment variables supplied to rustc.

This parameter emulated the impact of "cargo:rustc-env=VAR=VALUE"
lines, which are common in build scripts.