Use a macro to have only one rule exposed
This create a simple API for executing a build_script by just exposing
`cargo_build_script` which behaves like a rust_binary except creating
a

This creates an intermediate target `<name>_script_` that compiles
the binary and the final target is for executing the binary. This
could be joined as one rule by calling the `_rust_binary_impl` in
addition to the `_cargo_build_script_run`.