Correct CARGO_MANIFEST_DIR values for Bazel <3.7.0
Bazel 3.7.0 introduced a change (301b96b223b0c0) in the way it reports
file paths for external packages. This ensures that rules_rust sets the
right value for CARGO_MANIFEST_DIR both for Bazel < 3.7.0 and >= 3.7.0.

Note that if this changed once, it might change again, so the right
thing to do is to use a bazel API to do that work once it's available.