depenv: use the links attribute when set
The name of the environment variables present in the `.depenv` files for
sys packages should not be inferred from the package name but from the
`links` Cargo.toml attribute instead, like cargo does.

For backward-compatibility reasons, we fall back to inferring the
value from the name if the `links` attribute is unset.