Allow rustc env vars to be set in files
Build scripts are executed in the exec configuration, not the target
configuration, which means it's currently impossible to use values which
were generated as a result of the `--stamp` flag when creating a
`rust_binary`. This means that embedding things like build timestamps,
git shas, and dynamic version numbers, is impossible.

This change allows you to depend, in the target configuration (because
no transition occurs on the attribute) on the output of a genrule which
consumes data from the stable and volatile workspace status commands.