rustfmt_test: Find rustfmt manifests in manifest-style runfiles
`test_main.rs` was looking for runfiles by getting a path to the
runfiles dir, and then walking the tree looking for files with a
particular extension. This does not work where runfiles are
manifest-based rather than directory-based, as is the case on Windows.

This change fixes the issue by growing a `list_files` method on the
`Runfiles` struct that higher-level code can use to enumerate all
the runfiles, filtering out the files they're looking for.

Since `find_runfiles_dir` was public purely for this one use case, it
can become private; likely higher-level code shouldn't be using this
function directly, but rather interacting with runfiles through a
Runfiles object.

Issues: #619