Add CI tasks for rolling bazel version
This PR adds tasks for running our standard battery using current Bazel
rolling release. These tasks are only informative, they will not stop a
PR from submitting, but it's good to keep an eye on these and fix
quickly so we are prepared to incoming incompatible changes in Bazel.

This PR:

* adds Windows task
* adds Macos task
* adds Ubuntu 20.04 with Clang task
* removes Ubuntu 20.04 without aspects task (not necessary)