support native dependencies with linkstamps
This adds support for linking in native dependencies with linkstamps
into Rust binaries. The necessary Skylark features to support this were
added very recently to Bazel in
https://github.com/bazelbuild/bazel/commit/9894cba57b2247ac1d16854a1921a3cd4e6cdc2b,
which has only landed in recent Bazel 5.0.0 pre-release versions. This
implementation performs feature detection and only enables dependency
linkstamps if Bazel supports it.