Also propagate linkstamps through rust_libraries
Before this PR we were only propagating linkstamps from C++
dependency to Rust target, but we did not propagate from Rust to Rust.
This PR fixes it by moving the logic to collect linkstamps from
dependencies from the if branch that only covers C++ deps to a place
that affects all deps.