Provide `.dSYM` package in `OutputGroupInfo` for cdylib and lib on darwin
rustc emits symbols when the `split-debuginfo=packed` codegen option is
set (currently the default).

Also reorganized the tests a bit so that the logic between `pdb_file`
and `dsym_folder` is shared.

Note that while the cc rules use `pdb_file` in the `OutputGroupInfo` I
wasn't able to find anything similar for macos. So I opted for adding
some extra docs to `rust_shared_library` and `rust_binary` so that it's
easier to find.