add support for targeting iOS simulator on Apple Silicon
When building on an Apple Silicon machine, a different target triple
is required to target the simulator:
https://doc.rust-lang.org/rustc/platform-support/aarch64-apple-ios-sim.html

As rules_rust currently ignores abi in triples, listing device
and simulator triples in a rust_repository_set() ends up always using
the first one. Adding this constraint allows a build to be switched
between targeting a device and targeting the simulator.