add extra_exec_rustc_flags build config
This is analogous to
https://github.com/bazelbuild/rules_rust/commit/52e3ba4c99692b5354faff0fb1aa62ddf884024c
but for passing flags to the exec configuration. This is useful, e.g.,
when working in an environment similar to the stage0 bootstrapping one,
where the beta (stable) toolchain is used to compile unstable sources
with an additional `--cfg=bootstrap` flag to be passed everywhere in the
exec configuration.

One tricky thing was testing: I didn't find a good way to emulate the
testing for `extra_rustc_flags` from the above commit: checking that flags get
applied in exec configuration requires digging through the actions of transitive
dependencies of the target under test, and I didn't find a way to get to
those via the analysistest APIs. Instead I defined a chain of
transitions and a library rule on top that will only build if the
extra_exec_rustc_flags are indeed passed to exec transitions.