Remove the CARGO_BAZEL_ISOLATED env var.
Manipulating the build environment with this environment
variable doesn't seem super useful to me. Also, I am working
on a patch to allow isolation of the building of the cargo_bazel
binary when using a label as a generator instead of a generator_url.
This variable interferes with that patch since the
CARGO_BAZEL_ISOLATED env var doesn't really mean a lot for the
bootstrap case, so we would have to duplicate the cargo isolation
logic with slight differences if this variable is used in this
context. Duplicating the logic seems less useful to me than just
removing this var.