Add workaround for arm vs armv7 android issue
When building for the armv7 android target, we would ideally specify a
cpu:armv7 constraint, but because of the linked bazel issue this is
currently hardcoded as arm in bazel's NDK setup. As far as I can tell
the platform we request for use with rules rust has to match the
constraints in bazel, otherwise it won't discover the NDK CC toolchain
for linking.