Expand the universe of known shas to include the .xz archive shas.
I have explicitly added keys for the files with the extension explicitly
on the file key. The point of this is to (in a future change) allow
tar.xz archives to be used instead of tar.gz archives.

This change keeps compatiblity with the old keys that assume the tar.gz
archives. I don't know if this is something that we should drop in the future
or just not worry about.