Fix crate_universe/private/srcs.bzl to work with repo mappings
Currently the list in the said file only works if the rules_rust repo is named exactly "rules_rust". This commit fixes it so that the list contains Label objects which would work regardless of what the repo is called.

Also fixes the generation script to allow leading '@'s in the label stringification result (see https://github.com/bazelbuild/bazel/issues/16196).