Add example workspace cross-compiling with zig
We've run into a number of issues, both in rustc and in cargo build
scripts, where we're not properly propagating cc_toolchain information
to compiles and particularly link actions. These have often been hard to
test because we don't have an example set-up with a non-default C++
toolchain.

By wiring up a zig cross-compiling toolchain, we have a concrete example
we can work with. It's also an interesting potential example for folks
who want to set something similar up.