Use `_make_link_flags_darwin` when target os is `ios`.
Right now, `_make_link_flags_default` is used when the target os is `ios`.
However, it makes use of `--whole-archive` and `--no-whole-archive` flags,
which are not supported by `dyld`.

As it is done for the `macos` target, we should use `-force_load` instead.
`_make_link_flags_darwin` does that, therefore we should use it when linking
for iOS.