Detect patched paths
It looks like we have the info we need to tease out the
patched crates' paths without having to resort to completely parsing
the ID field (which is technically opaque). We can do a half measure
and look for `(path+file://` and then use other metadata in addition
to knowledge of the tempdir path length to form the correct crate
roots, and then make the globs/source_roots point to those which will
hopefully be enough to get them to build.

So far this change just patches up the source roots, and they're not
even correctly normalized as relative paths yet so it definitely won't
build yet. we may need to actually generate the build files in the
correct locations with an alias or something to make defs.bzl in the
vendor dir still point to it.