Handle corner case for windows architecture detection
It appears that `wmic os get osarchitecture` can also return

```
OSArchitecture
64-Bit
```

Note the capital `B`!

This fails for the string comparison and therefore fails to detect
the existing architecture for windows.

Unfortunately i could not find any documentation on why or when this
happens, but a simple `lower()` handles both cases just fine