Option to generate extra targets with crate_universe
This is similar to what is provided by cargo-raze's
`extra_aliased_targets`. As an example, we use this feature with the
cxx crate as the repo provides the headers needed in C++ sources. This
feature goes hand in hand with `additive_build_file` as we can create
the rule and generate the alias for that rule.