Add script to revendor all crates
It seems that vendored crates commonly get out of sync. Re-vendoring seems to currently be a quite manual process, and finding everything to re-vendor is a hassle.

This adds a bash script, `vendor.sh`, that finds all crate vendor rules in all workspaces in this repository and runs them. This significantly improves the workflow for updating these.