cargo_build_script: Set use_default_shell_env
Build scripts which use `cmake` often need to probe `$PATH`, e.g. on
macOS with the default C++ toolchain, `cmake` needs to probe `$PATH` to
find `/usr/bin/ar` as a default `ar` to use, because the default
toolchain uses `/usr/bin/libtool` (which the build script runner
explicitly filters out because it can't be used as a drop-in replacement
for `ar`).