Introduce feature `rules_rust_link_cc`.
This feature is similar to `rules_rust_unsupported_feature`. The
difference is that, `rules_rust_unsupported_feature` gets disabled if
*any* rust is involved, while this feature gets disabled if *only* rust
is involved.

My use case for this feature is to static link to c++ / objc stdlibs.
They are needed only when any C object is being linked, and will
produce invalid rust_proc_macro if linked blindly.

This implementation uses the fact that `DepInfo.transitive_noncrates`
would be empty unless there is a `cc_library` or `cc_import` in the
(transitive) deps. As a side effect, `collect_deps()` can no longer use
the output of `_are_linkstamps_supported()` so it now always returns a
populated linkstamp depset, to be discarded later if linkstamps are not
supported.