Use fake.rs not /dev/null as fake source file
We store in the lockfile the path of this source file relative to the
package, and if we use an absolute path we end up encoding the depth of
your local git clone into the lockfile, which is non-determinstic across
machines.

Elsewhere we use fake.rs, let's use it here as well.