fix!: correctly determine common deps and features
Currently, "common" deps and features in selects in generated lockfiles
are not always common to all target platforms in the presence of
conditional dependencies. This change ensures that common deps and
features for a crate are common to all supported targets, not just those
targets for which the crate dependency is enabled.