rust_analyzer: generate more reqired sources
The existing support for this has limitations:

- only works for sources that are specially tagged (rust_generated_srcs).
  I've removed this mechanism as it's no longer needed.
- only provides sources of the directly selected library, but
  rust-analyzer needs to parse its dependencies too.
- is missing the target.json file, which rust-analyzer requires to
  parse the library (it invokes rustc --print cfg --target=...)

The dependency depth to generate sources for is a tradeoff. More layers
of deps will give more accurate analysis but be slower to generate.

My guess is that in the end we'll want full transitive sources, but
direct deps should be a non-controversial starting point. (Without these
sources, rust-analyzer can't understand the direct API usage at all).
We can change this later and probably need to reconcile with #3028.