rust_analyzer: make all paths in rust-project.json absolute
This is a step towards supporting automatic project discovery (#2755).

Relative paths are resolved against the location of rust-project.json,
but in auto-discovery this is the location of the BUILD file being
discovered, not the workspace root.

We already use absolute paths for generated files, so rust-project.json
is not workspace-independent today.

The alternatives seem more complex for little gain:

- Only use absolute paths for auto-discovery mode, relative otherwise.
- Use ../../ to express all workspace paths relative to the BUILD.

See https://github.com/bazelbuild/rules_rust/issues/2755#issuecomment-250696107