Fix: Handle package.workspace key in child Cargo.toml
Setting the `package.workspace` key in child workspace fails splicing
with an error similar to

```
Found manifest at /home/mlaurin/src/check_mk.git/packages/site/check-http/Cargo.toml which is a member of the workspace at
/home/mlaurin/src/check_mk.git/packages/site/check-http/.. which isn't included in the crates_universe
```

Now, the key is valid in Cargo.toml when the code is organized in
workspaces.  It is even required for out-of-tree workspaces.  See
https://github.com/nox/rust-rfcs/blob/master/text/1525-cargo-workspace.md

Before December, the key was ignored but would not fail splicing.  The
regression is most likely due to changes related to
https://github.com/bazelbuild/rules_rust/issues/3090