Fix panic line numbers
Without this patch, line numbers are missing for first-party code in
panic traces.

This reverts part of #1563 but keeps the test showing we have a relative
path not an absolute one, and adds an integration test showing that line
numbers are correctly reported in panic traces.

As far as I know, path remapping has been improved in rustc over the
last few years, so maybe the behaviour reported in that patch has been
fixed.