Do not forward ARFLAGS (#3763)
This is a partial revert of #3704. It does NOT revert logic that
properly finds the hermetic archiver (binary and associated tests).

cc-rs needs full ownership of flag handling for the archiver. This is in
part due to `ar` having very specific positional flag behaviors, and
partially due to some special logic in how cc-rs constructs static
libraries.

The change to enable `ARFLAGS` was causing many toolchains that pass
flags to the static archiver to error out due to conflicting
expectations for flag handling. Example bad `ar` invocation, where `cq`
is treated as the destination archive instead of the `ar` tool mode:
```
  ar rcsD cq dest.a src1.o src2.o
```
This drops forwarding of `ARFLAGS` to fix the common case.