Fix generated script for vscode debug tasks
This is about the tool to debug rust targets in VS Code.

When generating the script used in `launch.json`, `stdout` is captured
instead of `stderr`. This results in an error when running the task in
VS Code (`IndexError: list index out of range`) then in another error
from lldb as no target has been registered.